fix(ac-607): restore solve designer retries
